[refactoring] create FileHandler class & move file method
